Only report single contradictions
Currently, GHC gives the binary choice of telling the solver that:

A) This one constraint out of the entire set is wrong
B) This subset of constraints can be solved

The problem with picking A is that the plugin will not be called
with the unsolved subset of constraints again; which leads to
constraints to could be solved by the plugin not being solved
at all, and hence errors.

So the plugin will only pick A is there are no solved constraint
and no remaining unsolved constraints; otherwise, when it finds
a contradiction it simply leaves it unsolved.