Implemented solving of `n <= Max (n + p) p` constraints, probably a few more constraints can also be solved now that couldn't be solved before.
The way this is achieved is now the solver keeps track of whether it has
normalised an expression within a constraint. If this plugin cannot solve
a constraint but it has normalised an expression within the constraint it
marks the original constraint as solved and emits a new partly normalised
constraint.